Make the rdsPattern format make more sense
Now, rely on `printf` in order to format and set the RDS strings. This
will hopefully make more sense for more people, and thankfully
disincludes the command itself from the configuration file, leaving
only the arguments. Quotes are necessary to avoid spaces interfering;
I was not able to get `crudini` option `--format=sh` to function
properly with this, but no matter, this works fine.

Advanced end-users can always edit getTitle() anyways. Perhaps this
should be in the README?